+++ packages/iov-bcp/types/transactions.d.ts
@@ -23,20 +23,19 @@ export declare function publicKeyBundleEquals(left: PublicKeyBundle, right: Publ
 /** Used to differentiate a blockchain. Should be alphanumeric or -_/ and unique */
 export declare type ChainId = string & As<"chain-id">;
 /** a public key we can identify with on a blockchain */
-export interface PublicIdentity {
+export interface Identity {
     readonly chainId: ChainId;
     readonly pubkey: PublicKeyBundle;
 }
-export declare function isPublicIdentity(data: any): data is PublicIdentity;
+export declare function isIdentity(data: any): data is Identity;
 /**
- * Compares two objects that conform to the PublicIdentity interface for equality.
- * This can also be used to compare pairs of derived types like LocalIdentity/PublicIdentity
- * or LocalIdentity/LocalIdentity in which case all non-PublicIdentity fields are ignored.
+ * Compares two objects that conform to the Identity interface for equality.
+ * All additional (non-Identity) fields are ignored.
  *
  * @param left the left hand side of the comparison
  * @param right the right hand side of the comparison
  */
-export declare function publicIdentityEquals(left: PublicIdentity, right: PublicIdentity): boolean;
+export declare function identityEquals(left: Identity, right: Identity): boolean;
 export declare type SignatureBytes = Uint8Array & As<"signature">;
 /** An integer in the safe integer range */
 export declare type Nonce = number & As<"nonce">;
@@ -131,7 +130,7 @@ export interface UnsignedTransaction {
      *
      * This implicitly fixes the chain ID this transaction can be used on.
      */
-    readonly creator: PublicIdentity;
+    readonly creator: Identity;
     readonly fee?: Fee;
 }
 export declare function isUnsignedTransaction(data: any): data is UnsignedTransaction;