+++ packages/iov-keycontrol/src/keyring.spec.ts
@@ -284,7 +284,7 @@ describe("Keyring", () => {
       expect(keyring.getWallets()[3]).toEqual(jasmine.any(Ed25519Wallet));
       expect(keyring.getWallets()[3].getIdentities().length).toEqual(4);
 
-      // compare wallet content (via PublicIdentity equality)
+      // compare wallet content (via Identity equality)
 
       expect(keyring.getWallets()[0].getIdentities()[0]).toEqual(i1a);