+++ packages/iov-keycontrol/types/wallet.d.ts
@@ -1,5 +1,5 @@
 import { As } from "type-tagger";
-import { ChainId, PrehashType, PublicIdentity, SignableBytes, SignatureBytes } from "@iov/bcp";
+import { ChainId, Identity, PrehashType, SignableBytes, SignatureBytes } from "@iov/bcp";
 import { Ed25519Keypair, Slip10RawIndex } from "@iov/crypto";
 import { ValueAndUpdates } from "@iov/stream";
 export declare type WalletId = string & As<"wallet-id">;
@@ -12,11 +12,11 @@ export interface ReadonlyWallet {
     /**
      * Gets a local label associated with the public identity to be displayed in the UI.
      */
-    readonly getIdentityLabel: (identity: PublicIdentity) => string | undefined;
+    readonly getIdentityLabel: (identity: Identity) => string | undefined;
     /**
      * Returns all identities currently registered
      */
-    readonly getIdentities: () => ReadonlyArray<PublicIdentity>;
+    readonly getIdentities: () => ReadonlyArray<Identity>;
     readonly canSign: ValueAndUpdates<boolean>;
     readonly implementationId: WalletImplementationIdString;
     /**
@@ -25,14 +25,14 @@ export interface ReadonlyWallet {
      * This allows the Keyring to check for duplicate identities before they
      * are persisted.
      */
-    readonly previewIdentity: (chainId: ChainId, options: Ed25519Keypair | ReadonlyArray<Slip10RawIndex> | number) => Promise<PublicIdentity>;
+    readonly previewIdentity: (chainId: ChainId, options: Ed25519Keypair | ReadonlyArray<Slip10RawIndex> | number) => Promise<Identity>;
     /**
      * Created a detached signature for the signable bytes
-     * with the private key that matches the given PublicIdentity.
+     * with the private key that matches the given Identity.
      *
-     * If a matching PublicIdentity is not present in this wallet, an error is thrown.
+     * If a matching Identity is not present in this wallet, an error is thrown.
      */
-    readonly createTransactionSignature: (identity: PublicIdentity, transactionBytes: SignableBytes, prehash: PrehashType) => Promise<SignatureBytes>;
+    readonly createTransactionSignature: (identity: Identity, transactionBytes: SignableBytes, prehash: PrehashType) => Promise<SignatureBytes>;
     /**
      * Exposes the secret data of this wallet in a printable format for
      * backup purposes.
@@ -65,10 +65,10 @@ export interface Wallet extends ReadonlyWallet {
      * The identity is bound to one chain ID to encourage using different
      * keypairs on different chains.
      */
-    readonly createIdentity: (chainId: ChainId, options: Ed25519Keypair | ReadonlyArray<Slip10RawIndex> | number) => Promise<PublicIdentity>;
+    readonly createIdentity: (chainId: ChainId, options: Ed25519Keypair | ReadonlyArray<Slip10RawIndex> | number) => Promise<Identity>;
     /**
      * Sets a local label associated with the public identity to be displayed in the UI.
      * To clear a label, set it to undefined
      */
-    readonly setIdentityLabel: (identity: PublicIdentity, label: string | undefined) => void;
+    readonly setIdentityLabel: (identity: Identity, label: string | undefined) => void;
 }